[AESS-46] Update agents_ranking_targets gsheet column casting (#8239)

+++ bietlejuice/jobs/composer/db/datalake/queries/gsheets/clean/support_and_service/agents_ranking_targets.sql
@@ -7,7 +7,7 @@ SELECT
     CAST(productivity_target AS INT) AS productivity_target,
     CAST(ra_would_do_business_again_target AS FLOAT) AS ra_would_do_business_again_target,
     CAST(ra_target_score AS INT) AS ra_target_score,
-    CAST(ra_target_solution_rate AS INT) AS ra_target_solution_rate,
+    CAST(ra_target_solution_rate AS FLOAT) AS ra_target_solution_rate,
     DATE(start_date) AS dt_start,
     DATE(end_date) AS dt_end
 FROM
